[arcane,core] Déplace la définition de 'ItemFlags' dans son propre fichier.
